fix(python): Fixed python import name from pychangepoint -> changepoint
This also removes the root `Cargo.toml` to prevent strange behavior wrt naming and to make it similar to `lace`.
