+++ .github/workflows/rust-build-test.yaml
@@ -21,6 +21,9 @@ env:
 jobs:
   features:
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: changepoint
     steps:
       - uses: actions/checkout@v3
 
@@ -154,6 +157,9 @@ jobs:
 
   release:
     name: release
+    defaults:
+      run:
+        working-directory: changepoint
     runs-on: ubuntu-latest
     if: startsWith(github.ref, 'refs/tags/rust')
     needs: ["semver-checks", "features", "lint", "test"]
@@ -162,6 +168,6 @@ jobs:
       - uses: dtolnay/rust-toolchain@stable
       - name: Publish to Crates.io
         run:
-          cargo publish --token ${CRATES_TOKEN} -p changepoint
+          cargo publish --token ${CRATES_TOKEN}
         env:
           CRATES_TOKEN: ${{ secrets.CRATES_TOKEN }}
