+++ pychangepoint/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "pychangepoint"
 description = "A Python module for changepoint"
-version = "0.2.0"
+version = "0.3.0"
 repository = "https://github.com/promised-ai/changepoint"
 authors = ["Redpoll <info@redpoll.ai>"]
 edition = "2021"
@@ -12,7 +12,7 @@ publish = false
 # prevents package from thinking it's in the workspace
 
 [lib]
-name = "pychangepoint"
+name = "changepoint"
 crate-type = ["cdylib"]
 
 [dependencies]
@@ -26,3 +26,6 @@ rv = { version = "0.14", features = ["process", "arraydist"] }
 [features]
 extension-module = ["pyo3/extension-module"]
 default = ["extension-module"]
+
+[package.metadata.maturin]
+name = "changepoint"
